requirements.txt
attrs==25.1.0
better-profanity==0.7.0
boto3==1.35.99
decord==0.6.0
diffusers==0.32.2
einops==0.8.1
huggingface-hub==0.29.2
hydra-core==1.3.2
imageio[pyav,ffmpeg]==2.37.0
iopath==0.1.10
ipdb==0.13.13
loguru==0.7.2
mediapy==1.2.2
megatron-core==0.10.0
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.535.133
omegaconf==2.3.0
opencv-python==4.10.0.84
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
protobuf==4.25.3
pynvml==12.0.0
pyyaml==6.0.2
retinaface-py==0.0.2
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==76.0.0
termcolor==2.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.49.0
